+++ src/Commands/CommandsHandler.lua
@@ -1,0 +1,18 @@
+--[[
+The commands handler provides resources for easy command registration,
+listening and triggering.
+]]
+local CommandsHandler = {}
+    CommandsHandler.__index = CommandsHandler
+    CommandsHandler.__ = self
+
+    --[[
+    Target constructor.
+    ]]
+    function CommandsHandler.__construct()
+        return setmetatable({}, CommandsHandler)
+    end
+-- end of CommandsHandler
+
+-- sets the unique library commands handler instance
+self.commands = CommandsHandler.__construct()