+++ src/stormwind-library.lua
@@ -1,5 +1,7 @@
 -- Library version = '0.0.7'
 
+-- import src/Commands/CommandsHandler.lua
+
 -- import src/Core/Factory.lua
 
 -- import src/Facades/Target.lua