+++ tests/Commands/CommandsHandlerTest.lua
@@ -1,0 +1,8 @@
+TestCommandsHandler = {}
+    -- @covers StormwindLibrary.commands
+    function TestCommandsHandler:testGetCommandsHandler()
+        local handler = __.commands
+
+        lu.assertIsTable(handler)
+    end
+-- end of TestTarget