+++ documentation/docs/changelog.md
@@ -2,6 +2,8 @@
 
 ## 2024.mm.yy - version 0.0.7-alpha
 
+* Add the CommandsHandler class to allow addons to register commands
+
 ## 2024.03.16 - version 0.0.6-alpha
 
 * Add the Str support class